Move intents convert to mapper commands + touchups
