+++ src/cmd/integrations/integrations.go
@@ -14,7 +14,7 @@ var IntegrationsCmd = &cobra.Command{
 	Use:     "integrations",
 	GroupID: groups.ResourcesGroup.ID,
 	Aliases: []string{"integration"},
-	Short:   "Integration operations via the Otterize API",
+	Short:   "Manage integrations via the Otterize API",
 }
 
 func init() {
