+++ src/cmd/intents/intents.go
@@ -2,7 +2,6 @@ package intents
 
 import (
 	"github.com/otterize/otterize-cli/src/cmd/groups"
-	"github.com/otterize/otterize-cli/src/cmd/intents/convert"
 	"github.com/otterize/otterize-cli/src/cmd/intents/get"
 	"github.com/otterize/otterize-cli/src/cmd/intents/list"
 	"github.com/spf13/cobra"
@@ -12,11 +11,10 @@ var IntentsCmd = &cobra.Command{
 	Use:     "intents",
 	GroupID: groups.ResourcesGroup.ID,
 	Aliases: []string{"intent"},
-	Short:   "Intent operations via the Otterize API",
+	Short:   "Manage intents via the Otterize API",
 }
 
 func init() {
-	IntentsCmd.AddCommand(convert.ConvertCmd)
 	IntentsCmd.AddCommand(get.GetCmd)
 	IntentsCmd.AddCommand(list.ListCmd)
 }
